material design finished

+++ src/app/pages/myevents/myevents.component.ts
@@ -1,14 +1,13 @@
-import { Component, OnInit, ViewChild, ViewContainerRef } from '@angular/core';
+import { ConfirmDialogComponent } from './../../components/confirm-dialog/confirm-dialog.component';
+import { Component, OnInit } from '@angular/core';
 import { User } from 'src/app/models/user';
 import { EventsService } from 'src/app/services/events.service';
 import { UsersService } from 'src/app/services/users.service';
 import * as moment from 'moment';
-
-
 import { getDataFromToken } from '../../utils/jwtparser';
 import { Event } from 'src/app/models/event';
-import { ErrorHandlerService } from 'src/app/services/error-handler.service';
 import { Subscription } from 'rxjs';
+import { MatDialog } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-myevents',
@@ -17,24 +16,14 @@ import { Subscription } from 'rxjs';
 })
 export class MyeventsComponent implements OnInit {
 
- 
-  image = "../../../assets/icons/my-event-icon.png";
-  star = "../../../assets/icons/big-star.png";
-  trash = "../../../assets/icons/trash-icon.png";
-  editIcon = "../../../assets/icons/pencil-icon.png";
-  plus = "../../../assets/icons/plus-icon-empty.png";
-  miniStar = "../../../assets/icons/mini-star.png";
-  
-
   userEmail: string = getDataFromToken().username;
   userPosition: number = 0;
   user: User = new User();
   pastEvents: Event[] = [];
-  futureEvents: Event[] ;
+  futureEvents: Event[];
 
-  ErrorMessage:string;
 
-  constructor(private userService: UsersService, private eventService: EventsService,  private errorHandlerService:ErrorHandlerService) { }
+  constructor(private userService: UsersService, private eventService: EventsService, public dialog: MatDialog) { }
 
   ngOnInit(): void {
     this.getUserByEmail();
@@ -46,22 +35,12 @@ export class MyeventsComponent implements OnInit {
   getUserByEmail() {
     this.userService.getUserByEmail(this.userEmail).subscribe((user) => {
       this.user = user;
-    },
-    (error) => {
-      this.ErrorMessage=error.error.message;
-      this.createModal();
-
     })
   }
 
   getUserPosition() {
     this.userService.getUserPosition(this.userEmail).subscribe((position) => {
       this.userPosition = position;
-    },
-    (error) => {
-      this.ErrorMessage=error.error.message;
-      this.createModal();
-
     })
   }
 
@@ -70,12 +49,6 @@ export class MyeventsComponent implements OnInit {
       return this.futureEvents = events.filter((event) => {
         return moment(event.date).isAfter(moment());
       })
-
-    },
-    (error) => {
-      this.ErrorMessage=error.error.message;
-      this.createModal();
-
     });
   }
 
@@ -84,40 +57,29 @@ export class MyeventsComponent implements OnInit {
       return this.pastEvents = events.filter((event) => {
         return moment(event.date).isBefore(moment());
       });
-    },
-    (error) => {
-      this.ErrorMessage=error.error.message;
-      this.createModal();
-
     });
   }
-//
-  deleteEventById(id:number){
-    this.eventService.deleteEventById(id).subscribe(
-      res=>{
-        return this.getFutureEventsByUser(); },
-    (error) => {
-      this.ErrorMessage=error.error.message;
-      this.createModal();
+  
 
+  deleteEventById(id: number, title: string) {
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      width: '250px',
+      height: '300px',
+      data: { title: "Eliminar evento", message: "¿Desea eliminar el evento " + title + "?" },
     });
-  }
-    
-  formatDate(date:Date){
-    return moment(date).format("DD-MM-YYYY");
-  }
 
-    //Error handler modals
-    @ViewChild('modal', { read: ViewContainerRef })
-    entry!: ViewContainerRef;
-    sub!: Subscription;
-  
-  
-    createModal(){
-        this.sub = this.errorHandlerService
-          .openModal(this.entry, 'ERROR', this.ErrorMessage)
-          .subscribe((v) => {
-            //your logic
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.eventService.deleteEventById(id).subscribe(
+          res => {
+            return this.getFutureEventsByUser();
           });
-    }
+      }
+    })
+}
+
+formatDate(date: Date){
+  return moment(date).format("DD-MM-YYYY");
+}
+
 }
